+++ unity-2017/Assets/Scripts/Database/DarkestJsonReader.cs
@@ -6,37 +6,6 @@ using Newtonsoft.Json;
 
 namespace DarkestJson
 {
-    #region Heirloom Trading
-    public class JsonHeirloomExchange
-    {
-        public List<JsonHeirLoomMarket> markets;
-    }
-    public class JsonHeirLoomMarket
-    {
-        public string id;
-        public List<JsonHeirloomExchangeEntry> exchange_rates;
-    }
-    public class JsonHeirloomExchangeEntry
-    {
-        public string exchange_from_type;
-        public int exchange_from_amount;
-        public string exchange_to_type;
-        public int exchange_to_amount;
-    }
-    #endregion
-
-    #region Party Names
-    public class JsonPartyNameDictionary
-    {
-        public List<JsonPartyNameEntry> party_names;
-    }
-    public class JsonPartyNameEntry
-    {
-        public string id;
-        public List<string> required_hero_class;
-    }
-    #endregion
-
     #region Narration
     public class JsonNarration
     {
@@ -869,16 +838,6 @@ namespace DarkestJson
         public static JsonClass GetJsonObject<JsonClass>(string jsonString)
         {
             return JsonConvert.DeserializeObject<JsonClass>(jsonString);
-        }
-
-        public static JsonHeirloomExchange GetJsonHeirloomExchange(string exchangeString)
-        {
-            return JsonConvert.DeserializeObject<JsonHeirloomExchange>(exchangeString);
-        }
-
-        public static JsonPartyNameDictionary GetJsonPartyNames(string partyString)
-        {
-            return JsonConvert.DeserializeObject<JsonPartyNameDictionary>(partyString);
         }
 
         public static JsonNarration GetJsonNarration(string narrationString)