+++ unity/Assets/Scripts/Campaign/HeirloomExchange.cs
@@ -1,7 +1,0 @@
-﻿public class HeirloomExchange
-{
-    public string FromType { get; set; }
-    public int FromAmount { get; set; }
-    public string ToType { get; set; }
-    public int ToAmount { get; set; }
-}